tests: Fix warning due to changed signedness of datum
  "test/t-byte-buffer.c", line 153:
      warning: integer conversion resulted in a change of sign

These warnings are emitted by TI's legacy toolchains.

+++ test/t-byte-buffer.c
@@ -150,7 +150,7 @@ main(UNUSED int argc, UNUSED char **argv)
         byte_buffer_clear(&s);
         byte_buffer_fillx(&s, 0x10u, -3);
         for (size_t i = 0u; i < m; ++i) {
-            buf[i] = i*-3 + 0x10u;
+            buf[i] = (ssize_t)i * (-3) + 0x10u;
         }
         cmp_mem(s.data, buf, m,
                 "_fillx fills buffer correctly (neg. increment)");
